Swap to using go embed and update page switching
Go Embed lets us pull in files as strings and more at compile time
so we only need to ship one binary with everything bundled in.

Page switching was swapped so that we have just 2 main pages, the intro
and the content. Then we move the rules and other bits of content into
their own page and have that in the flex for layout and toggle switching that
rather than the whole outer page. This ensures focus stays on the menu but can
be moved into the nested page if required (for character etc eventually).